update 'Database classes', 'DummyContent' and 'Database Helper'

+++ app/src/main/java/com/example/habitsbuilder/Database/Achievements.java
@@ -3,16 +3,16 @@ package com.example.habitsbuilder.Database;
 import java.io.Serializable;
 
 public class Achievements implements Serializable {
-    private long AchieveId;
+    private int AchieveId;
     private String Name;
     private String Description;
-    private long RewardID;
+    private int RewardID;
     private String Image;
     private int State;
 
     public Achievements(){};
 
-    public Achievements(String name, String des, long rewId, String img, int state){
+    public Achievements(String name, String des, int rewId, String img, int state){
         Name = name;
         Description = des;
         RewardID = rewId;
@@ -20,17 +20,17 @@ public class Achievements implements Serializable {
         State = state;
     }
 
-    public long GetAchId(){return AchieveId;}
+    public int GetAchId(){return AchieveId;}
     public String GetAchName(){return Name;}
     public String GetAchDes(){return Description;}
-    public long GetAchRewId(){return RewardID;}
+    public int GetAchRewId(){return RewardID;}
     public String GetAchImg(){return Image;}
     public int GetAchState(){return State;}
 
-    public void SetAchId(long id){AchieveId = id;}
+    public void SetAchId(int id){AchieveId = id;}
     public void SetAchName(String name){Name = name;}
     public void SetAchDes(String des){Description = des;}
-    public void SetAchRewId(long rewId){RewardID = rewId;}
+    public void SetAchRewId(int rewId){RewardID = rewId;}
     public void SetAchImg(String img){Image = img;}
     public void SetAchState(int state){State = state;}
 }
