+++ app/src/main/java/com/example/habitsbuilder/Database/HabitDay.java
@@ -4,7 +4,7 @@ import java.io.Serializable;
 import java.util.Date;
 
 public class HabitDay implements Serializable {
-    private long HabitId;
+    private int HabitId;
     private Date Date;
     private int State;
 
@@ -15,11 +15,11 @@ public class HabitDay implements Serializable {
         State = state;
     }
 
-    public long getHabitId(){return HabitId;}
+    public int getHabitId(){return HabitId;}
     public Date getDate(){return Date;}
     public int getState(){return State;}
 
-    public void setHabitId(long habitId){HabitId = habitId;}
+    public void setHabitId(int habitId){HabitId = habitId;}
     public void setDate(Date date){Date = date;}
     public void setState(int state){State = state;}
 }
